uninstate and reinstate tests for practical_examples
Some notebooks need updating or of some the downloaders
don't work. For the ones that are fixed the tests are
reinstated or remain working. For the ones that are
too much work for this moment, the tests are commited out.
Issues are made in the Github repository so we pick this
up soon.